requirements-mcp.txt
mcp>=1.0.0
websocket-client>=1.7.0
